requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
dill==0.4.0
docling==2.43.0
docling-core==2.44.1
docling-ibm-models==3.9.0
docling-parse==4.1.0
easyocr==1.7.2
et_xmlfile==2.0.0
Faker==37.6.0
fastapi==0.116.2
filelock==3.18.0
filetype==1.2.0
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
google==3.0.0
google-auth==2.40.3
google-genai==1.36.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
icrawler==0.6.10
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.10
latex2mathml==3.78.0
lazy_loader==0.4
lxml==5.4.0
markdown-it-py==3.0.0
marko==2.1.4
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pdf2image==1.17.0
pdfkit==1.0.0
pillow==11.3.0
pluggy==1.6.0
polars==1.33.1
propcache==0.3.2
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.26.4
pyparsing==3.2.3
pypdfium2==4.30.0
PySocks==1.7.1
python-barcode==0.16.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.1
selenium==4.35.0
semchunk==2.2.2
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.4
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.199
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.36.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xlsxwriter==3.2.5
yarl==1.20.1
zstandard==0.23.0
gunicorn==23.0.0


